kestra/pyproject.toml
[project]
name = "santander-cycles-pipeline"
version = "1.0.0"
description = "End-to-end batch data pipeline for TfL Santander Cycles usage data"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.13",
    "google-cloud-storage>=2.18",
    "requests>=2.31",
    "polars>=1.39.3",
    "streamlit>=1.55.0",
    "plotly>=6.6.0",
    "duckdb>=1.5.1",
    "fastexcel>=0.19.0",
]
